refactor: utils and new loading style (#276)
