+++ packages/client/src/stores/useMessage.ts
@@ -5,6 +5,7 @@ import { defineStore } from 'pinia'
 import { ref } from 'vue'
 import { toast } from 'vue-sonner'
 
+import { createMediaBlob } from '../utils/blob'
 import { useSettingsStore } from './useSettings'
 import { useWebsocketStore } from './useWebsocket'
 
@@ -28,39 +29,7 @@ export const useMessageStore = defineStore('message', () => {
 
       const chatMap = useMessageChatMap(chatId)
 
-      function getMediaMimeType(mediaType: string): string {
-        switch (mediaType) {
-          case 'photo':
-            return 'image/jpeg'
-          case 'sticker':
-            // Telegram stickers are usually WebM or WebP
-            return 'video/webm'
-          case 'document':
-            return 'application/octet-stream'
-          default:
-            return 'application/octet-stream'
-        }
-      }
-
-      for (const media of message.media ?? []) {
-        if (media.byte) {
-          const buffer = new Uint8Array((media.byte as any).data)
-
-          const mimeType = getMediaMimeType(media.type)
-          const blob = new Blob([buffer], { type: mimeType })
-          const url = URL.createObjectURL(blob)
-          media.blobUrl = url
-
-          // eslint-disable-next-line no-console
-          console.log('[MessageStore] Blob URL created:', {
-            url,
-            mimeType,
-            blobSize: blob.size,
-          })
-
-          media.byte = undefined
-        }
-      }
+      message.media = message.media?.map(createMediaBlob)
 
       chatMap.set(message.platformMessageId, message)
     })
@@ -71,7 +40,7 @@ export const useMessageStore = defineStore('message', () => {
       let restMessageLength = pagination.limit
       const dbMessages: CoreMessage[] = []
 
-      if (!useSettingsStore().messageDebugMode) {
+      if (useSettingsStore().useCachedMessage) {
         websocketStore.sendEvent('storage:fetch:messages', { chatId, pagination })
         const { messages: dbMessages } = await websocketStore.waitForEvent('storage:messages')
