+++ packages/client/src/stores/useSettings.ts
@@ -17,7 +17,7 @@ export type ChatGroup = DialogType | ''
 
 export const useSettingsStore = defineStore('settings', () => {
   const selectedGroup = useLocalStorage<ChatGroup>('settings/group-selected', 'user')
-  const messageDebugMode = useLocalStorage<boolean>('settings/message-debug-mode', false)
+  const useCachedMessage = useLocalStorage<boolean>('settings/use-cached-message', false)
 
   const theme = useLocalStorage<string>('settings/theme', 'default')
   const storageConfig = useLocalStorage<Config>('settings/config', generateDefaultConfig())
@@ -62,6 +62,6 @@ export const useSettingsStore = defineStore('settings', () => {
     applyPrimaryColorFrom,
     config: storageConfig,
     selectedGroup,
-    messageDebugMode,
+    useCachedMessage,
   }
 })
