+++ packages/stage-ui/src/components/messages/media/MediaRenderer.vue
@@ -87,9 +87,8 @@ const finalError = computed(() => {
   </div>
 
   <!-- Loading state -->
-  <div v-if="isLoading" class="flex items-center gap-2">
-    <div class="i-lucide-loader-circle h-4 w-4 animate-spin" />
-    <span class="text-xs text-complementary-600">处理媒体中...</span>
+  <div v-if="isLoading" class="h-20 w-20 flex items-center justify-center">
+    <div class="h-20 w-20 animate-pulse rounded-lg bg-gray-200 dark:bg-gray-700" />
   </div>
 
   <!-- Error state -->
@@ -117,7 +116,7 @@ const finalError = computed(() => {
     <video
       v-else-if="processedMedia.type === 'sticker'"
       :src="processedMedia.src"
-      class="h-auto max-w-xs rounded-lg"
+      class="h-auto max-w-[12rem] rounded-lg"
       alt="Media content"
       autoplay loop muted playsinline
       @error="runtimeError = 'Sticker failed to load'"
